WT-14137 Atomically copy "regular" files on open and remove the log copy stage of live restore. (#11575)
The log copy stage was added because tracking extents on log files and
similar things seemed ugly, and for a while that was valid. During the
development of the bitmap change it became clear that a simpler path
existed: treat all "regular", i.e. log files and non b-tree files as
copy-on-open. That way only b-trees have hole lists that need tracking.

Furthermore to avoid wondering the what the impact of a crash during
copy of these files would be the copy has been made atomic by doing a
copy followed by rename.

---------

Co-authored-by: Andrew Morton <andrew.morton@mongodb.com>
Co-authored-by: Sean <41533874+Sean04@users.noreply.github.com>