WT-13736 Fix test_schema_abort generating out of order oldest timestamp (#11581)
This pull request has the changes to fix a bug in the test_schema_abort
where two threads incorrectly compares last_oldest_timestamp with
current_oldest_timestamp and proceeds to call the set_timestamp API if
the difference exceeds 100. However, this difference can sometimes be
negative, leading to unintended calls to the set_timestamp API and
triggering the observed failure.