WT-14040 Implement live restore state tracking locking and correctness (#11590)
Add more granular states to Live Restore for each of the stages it
progresses through. This change adds checks that we always progress
through the changes in the correct order and adds both validation and
behaviour changes based on the current state WiredTiger is in.

The state is stored in memory during operation, but also persisted to
disk to allow WiredTiger to restart in the correct state across restarts
and crashes.