WT-13979 Fix wt dump so that it exports valid json even when searching for a nonexistent key (#11585)
The cause of the issue is that, at the point where the wt utility
detects during the dump process that a key is not found (ie doesn't
exist), the wt utility immediately exits. When outputting plain text
this approach works. However, when outputting as json, additional output
is still required to close any open square brackets and/or braces.

The Python test test_dump04.py already tested that the wt utility
handled json output when asked for a non-existent key. However, the test
did not check that the resulting json was actually valid - adding that
check to the test without fixing the bug finds this issue immediately.

The changes in the PR:
- Modify test_dump04.py to parse the json output and generate an error
if it is invalid.
- Modify util_dump when outputting json to correctly output 'end of
data' square brackets and 'end of table' braces in both the 'existing
key' and 'non-existent' key cases.
- Updated the command line (-?) docs to explain than no error is
generated when searching for a non-existent key (as before, the tool
simply displays no data).

Unfortunately, the util_dump function has a very high cyclomatic
complexity, and this change increases it by 1. I attempted to find small
changes to the function to reduce the complexity to compensate, but that
proved non-trivial due to the tight coupling within this large function.
As a result, I have created a seperate ticket to tackle that.