WT-14160 Remove unused sub-level error code WT_SESSION_MAX (#11608)
The WT_SESSION_MAX is removed from the project. The reason behind is
because the session API structure does not track the number of sessions
opened at once. This will be needed to be done on the connection level
API, but the function is not supported on the connection level.