SLS-1276 Include metadata of newly created tables in the next checkpoint (#11587)
Make sure that empty tables get included in the next checkpoint, so that
they would be present at the follower that picks up the checkpoint, or
at a new leader after restarting without local files. The PR does this
by keeping a list of URIs of newly created tables, which would get their
metadata copied to the shared memory table at the next checkpoint.

I tried to first update the shared metadata table during schema create,
but that worked only for `layered:` tables, but not for `type=layered`,
as `colgroup:` was not available at the time of layered table creation.
And it also had complicated error handling, because updating the shared
metadata table is not currently part of metadata tracking.

I also tried marking the shared component of a new layered table to be
included unconditionally in the next checkpoint, but it turns out that
this requires opening a dhandle during schema create. This is
problematic, because (1) it can cause a deadlock in some circumstances,
and (2) it does not account for the possibility that the dhandle could
be swept before the next checkpoint (although it would be unlikely due
to the timing).

The PR also moves some of the code for updating the shared metadata
table to `conn_layered.c`, so that it is together with the code that
reads the shared metadata table.