WT-13283 Cache aggressive mode shows garbage value on 6.0 FTDC (#11622)
We currently do check before decrementing `evict_aggressive_score` to
ensure it won't drop below zero. However the check then decrement
operations are not implemented atomically which causes race condition.
This PR is to apply the compare-and-swap operation to the code to avoid
the value drops to -1(which is `int_max` as `evict_aggressive_score` is
an unsigned int).

(cherry picked from commit 4a0a6109f05468cdfb1a7964e98bcfb4c351aa88)