WT-14199 Remove log prints in filesystem calls (#11656)
The `__wt_log_printf` in `__posix_fs_rename` was the culprit of all the
deadlocks seen in evergreen. I've removed this and the other calls to
`__wt_log_printf` in `os_fs.c` as they create similar dependencies. Let
me know if you think we should keep the others in `open_file` and
`remove`. Alternatively, if we want to keep all of them, we can add
```
 if (!F_ISSET(conn, WT_CONN_LIVE_RESTORE_FS))
```
and only print them if we're not in live restore mode.

The deadlock does not reproduce with these removed.