WT-14233 Reduce the amount of state persisted on the live restore file handle. (#11677)
The live restore file handle tracks a number of fields, two of them:
complete and source_size can be determined using other information. I've
changed complete to simply mean the absence of a source file handle and
source_size is no longer used. The canonical size is therefore the size
of the destination file on disk.