WT-13971 Return ENOENT when live restore attempts to open a directory that doesn't exist in the destination (#11681)
This change addresses both WT-13971 and WT-13823. 
As we're only opening and migrating file WiredTiger owns we can keep the
logic simple. WiredTiger never creates directories so any directory we
open must already exist, and files use the default permissions as set in
the posix or windows layer which live restore calls into.