WT-14111 Correct the documentation for forced eviction behaviour (#11694)
Update the `arch-eviction.dox` to more accurately describe the
conditions when pages are added to the urgent eviction queue. Pages will
only go into the urgent queue if the operation has disabled eviction or
splitting, or if the session is preventing reconciliation. Otherwise the
page can be evicted by the application thread itself.