WT-14243 Fix deadlock between the turtle lock and the live restore state lock (#11695)
This change adds two functions to intercept calls to update or rewrite
the turtle file. Ensuring that the live restore state file lock is held
before the turtle lock is taken. The consistent order of claiming the
two locks removes the possibility of a dead lock when one thread set the
live restore at the same time another thread is updating the turtle
file.