WT-14195 Reduce set_last_error calls in API_INIT (#11666)
As part of [PM-3810](https://jira.mongodb.org/browse/SPM-3810) and
[WT-13866](https://jira.mongodb.org/browse/WT-13866), we have introduced
a performance regression. The performance regression is due to the
number of calls calling set last error within each session API call.
Currently, we call the set_last_error function to return the session
error info structure to the default state before performing the actual
session API work. This was done because of the need to have error state
clean before we start tracking errors within WiredTiger.

**Solution**
The solution to this problem is to only perform set_last_error if the
previous session error info contains a fail error code. In this case it
means that we do not need to perform unnecessary resetting to default
and only set it back to default if any of the members inside the error
info structure was changed.

This also requires change in the API behaviour. After a success full
open session call, the error message will reflect the successful open of
a session with a success message. Previously it would be an empty
string.