WT-13716 Prevent retention of fast truncated history store pages (v8.0)(#11190) (#11710)
The problem described in the ticket is that, since the HS is
non-transactional, page_del->committed is not set. This prevents the
deletion of the page, causing reconciliation to retain the original
version. By treating HS pages that are fast-truncated during selective
backup as globally visible (even though they are non-globally visible)
and setting page_del to null, we eliminate this issue. This allows us to
remove the pages without checking visibility information, which is
acceptable because the HS is non-transactional.

(cherry picked from commit 472b1d8fd675b1453eecf07642dd88bbf2ed79fa)

Summarize the reason behind this change (this might be the problem
you're solving, or the context around the request) and the solution you
have chosen.

---------

Co-authored-by: Siddhartha Mahajan <siddhartha.mahajan8899@gmail.com>