WT-14257 Remove Windows ifdefs in the live restore code (#11707)
Previously we needed to #ifdef out live restore code for Windows as live
restore used posix-specific libraries such as ioctl fiemap. Now that we
persist live restore data to the metadata and turtle files this is no
longer necessary and we can clean up the code.

This change only removes pre-processor #ifdefs, live restore is still
not supported on Windows.