WT-13474 Remove csuite-long-running from the ubuntu2004-msan variant (#11717)
The csuite-long-running task frequently times out on the
csuite-long-running variant, and looking through Jira all of our tickets
relating to this task are about timeouts and not MSan detecting errors.
Rather than increase the timeouts once more this ticket deletes the
task. We still run the task under MSan on the amazon2023-armv9-msan
variant, so we get the same sanitizer coverage in half the runtime.