WT-13283 Fix the cache aggressive mode not to show the garbage value (8.0) (#11622) (#11711)
We currently do check before decrementing `evict_aggressive_score` to
ensure it won't drop below zero. However the check then decrement
operations are not implemented atomically which causes race condition.
This PR is to apply the compare-and-swap operation to the code to avoid
the value drops to -1(which is `int_max` as `evict_aggressive_score` is
an unsigned int).

(cherry picked from commit 4a0a6109f05468cdfb1a7964e98bcfb4c351aa88)


Co-authored-by: Zunyi Liu <43444919+zunyiliu@users.noreply.github.com>