WT-10234 Don't run code with side effects inside asserts (#11720)
Code inside a WT_ASSERT will only run when we compile in HAVE_DIAGNOSTIC
mode. In release mode it's entirely compiled out. This carries a risk
when the code inside the assert has some side effect, for example
setting a field in a struct.
This ticket removes the few cases where we do this. They're already
inside of an #ifdef HAVE_DIAGNOSTIC so there's no functional impact, but
it protects against the possibility of ever using this code outside of
HAVE_DIAGNOSTIC.