WT-14212 Fix test flakiness due to low timer granularity (#11722)
Fix test flakiness due to low timer granularity.

The purpose of this test is to verify the eviction interruption
mechanism. It relies on certain metrics, and as a side effect, it also
ensures the correctness of those metrics.

However, on systems with low clock granularity, this part of the test is
unreliable. One way to mitigate this would be to run the test longer,
but that would increase runtime on all platforms.

Instead, I reduced the metrics coverage in this test. Metrics can be
tested separately in a dedicated test.