WT-14177 Don't allow live restore and salvage to be configured at the same time (#11726)
In rare cases during salvage WiredTiger will delete a corrupted turtle
file and recreate it using the metadata file. This presents an issue as
the live restore global state is only stored in the turtle and we would
lose that data.
To prevent this don't allow salvage and live restore to run at the same
time.