WT-14065 Add additional performance statistics to live restore (#11727)
Adding a latency histogram in milliseconds because anything less than a
millisecond is not interesting for source reads.
Adding a statistic for how many bytes are being copied by the background
thread.
Adding a number of source file reads stat which should gradually
decrease as the live restore continues.

I am currently testing in a patch build and will post FTDC to the ic
channel prior to merge.