WT-14235 Atomically set the destination file size (#11733)
The destination file size is the cannonical file size but we have no
guarantee we won't crash between creating a file in the destination and
setting its size. This change ensures that the file creation and file
size setting are atomic by renaming a temporary file.

I did manage to reproduce this failure using failpoints and some code
editing but I don't see the point committing a failpoint that will only
enable a single test.

Interestingly the following check was disabled to get my test to pass: 
```
        // for (uint32_t i = 0; i < num_dest_files; ++i) {
        //     if (WT_PREFIX_MATCH(dirlist_dest[i], WT_WIREDTIGER) ||
        //       WT_SUFFIX_MATCH(dirlist_dest[i], ".wt"))
        //         WT_ERR_MSG(session, EINVAL,
        //           "Attempting to begin a live restore on a directory that already contains "
        //           "WiredTiger files '%s'! It's possible this file will be overwritten.",
        //           dirlist_dest[i]);
        // }
```
We should implement more crash testing where we can.

---------

Co-authored-by: Andrew Morton <andrew.morton@mongodb.com>
Co-authored-by: Sean <41533874+Sean04@users.noreply.github.com>