WT-13161 Pass cache_size into __cache_config_abs_to_pct as an argument (#11730)
We've been caught in previous PRs (see the Jira ticket) by moving
`__evict_config_abs_to_pct` earlier in the code to a point before we set
`conn->cache_size`, resulting in incorrect configurations. This is
because the dependency of `__evict_config_abs_to_pct` on
`conn->cache_size` is not obvious.

By passing `conn->cache_size` as an argument instead of fetching it from
the connection, `__evict_config_abs_to_pct` becomes a pure function
which makes it easier to reason about the value returned and harder to
accidentally call it before `conn->cache` has been configured.