WT-14278 Clean up live restore bitmap metadata on backup open, not on backup close (#11757)
This change simplifies the lifecycle of the live_restore=(nbits)
metadata item.
When a file finished migration we set nbits=-1, but when the entire live
restore process completed we would then transition to nbits=0 in
preparation for the current database to be used as a source in future
live restore.

However we could only set nbits=0 if the file was opened and then
dirtied so it wrote metadata back to disk. This lead to a lot of
scenarios where a fully migrated file sometimes has nbits=0, and
sometimes had nbits=-1 even though they both meant the same thing.

To simplify this logic we no longer set nbits=0 during normal
operations. nbits=-1 indicates that the migration has completed, and we
only clean the metadata to say nbits=0 during the process of taking a
backup.

This means that only a valid backup can be used as a source directory,
so this change adds that condition and updates all of our tests to use
valid backups in their source directories.