WT-14151 Don't clobber existing directories in cppsuite's connection_manager::create (#11772)
I ran into this a few times when implementing a prior live restore test.
connection_manager::create will silently delete any existing folder of
the same name, so when a destination folder hasn't been copied across to
the source it makes it harder to diagnose what went wrong when the
eventually crashes. Add a quick safety check to prevent this from
happening in future.