WT-14311 Add heavy testing for live restore (#11774)
Turn on additional testing for live restore for the final weeks of the
project.
We're writing additional unit tests for live restore, but to help
improve confidence in code correctness bump up the amount stress testing
for live restore. This change adds approximately 10 hours of additional
live restore testing per commit in waterfall.

This increase in testing is temporary. These tests will be removed again
at the end of the project to save on compute.