WT-14315 Disable live restore on 8.1 (#11784)
The 8.1 branch contains a partially implemented version of live restore
which also contains bugs fixed on develop. We won't be backporting live
restore to 8.1 so this change both disallows the config - returning
ENOTSUP when live restore is provided - and deletes most of the live
restore tests.
I've left test/format alone as the config has a zero percent chance of
firing in random executions.