WT-14068 Fix misaligned usage of modify variable (#11781)
**Problem statement**
A UBSAN failure has been found when needing to loop through a modify.
All modify data is captured within the `WT_UPDATE->data` variable. The
first `size_t` bytes of the modify refer to the number of entries within
one modify. Followingly, the `WT_MODIFY_FOREACH_BEGIN` and
`WT_MODIFY_FOREACH_REVERSE` loops through the same pointer and captures
the metadata of a modify fetching data size, offset ,and size, all of
which are `size_t` bytes.

The `WT_UPDATE` structure is packed such that `WT_UPDATE->data` always
starts from the 47th byte. In the general case, this is safe. However,
when dealing with modifies, the codebase typecasts the `WT_UPDATE->data`
structure to **size_t* **. This is problematic as we are starting from
an unaligned position (47 % 8 != 0).

**Solution**
The ticket will address the misaligned behaviour by fetching the size_t
information under a `uint8_t` pointer without typecasting it. While
addressing this, I encountered further improvements:
- The `napplied` variable is not needed inside `WT_MODIFY_FOREACH_BEGIN`
- Changed `nentries` and `napplied` to size_t. We need to ensure they
can never be negative.
- Removed the `tmp` variable, as it is not required.