WT-14325 Skip syscall test if env issues detected
Currently running simple `ctest` command on a non-preconfigured macOS system reports syscall test failure because
of Apple SIP system (enabled by default since OSX El Capitan) which restrict tracing system calls using `dtrace` tool.
The idea of this patch is to mark this test as "skipped" in this case instead of "failed" since generally
speaking it is not a test failure (like WT calling unexpected system call) but rather an absence of testing due to
lack of required permissions. This commit also shows a message about the reason of the test being skipped.