WT-14327 Evaluate __wt_verbose_multi_* macro arguments once (#11799)
We often pass `WT_VERB_RECOVERY_RTS` into the `__wt_verbose_multi_*`
macros as the `multi_category` argument, but `WT_VERB_RECOVERY_RTS` is a
ternary under the hood so when we substitute `multi_category` into the
macro we've evaluating the ternary in three separate locations.

clang-analyzer raises a warning about this, as when
`WT_VERB_RECOVERY_RTS` evaluates differently in different locations we
can attempt a read past the end of the array. In the following code
```c
for (__v_idx = 0; __v_idx < multi_category.cnt; __v_idx++) {
    if (WT_VERBOSE_LEVEL_ISSET(session, multi_category.categories[__v_idx], level)) {
```
if `multi_category` first evaluates to a struct with two categories
(`cnt == 2`) then `__v_idx` will iterate up to a value of 1. Then if
`multi_category` evaluates to a single-category struct in the second
occurrence then `multi_category.categories` is an array of length one
and `__v_idx == 1` attempts to read past the end of the array.

The solution is to evaluate `multi_category` once and store the result
into a local variable.