SLS-1397 Skip writing empty delta (#11788)
This change uses a different approach.

In the previous approach, I used an internal error to interrupt the
reconciliation flow if we skip writing an empty delta. However, it is a
hack and it generates a fallout on mongod due to an unknown problem.

In the new approach, what I do is to copy the previous written disk
address as if it is written this time when we skip writing an empty
delta. There are three cases:
* We copy the previous disk address from mod_replace.
* We copy the previous disk address from mod multi structure.
* We fall back to the existing ref->addr.