WT-14119 Make sure all static functions in live restore check their return values (#11802)
WiredTiger annotates almost all of its functions with
`WT_GCC_FUNC_DECL_ATTRIBUTE((warn_unused_result))` to make sure the
compiler warns us when a function returns a value but we don't use or
check that value.
However, we can't add this attribute to static functions and don't get
warnings for mis-use of these functions.
This change is a manual review of all static functions in
`src/live_restore` to properly check error returns for non-void static
functions.