WT-14268 Disable table_logging for prepare_stress.py (#11806)
The prepare-stress workload generates 250 threads altogether to perform
CRUD operations. The *table_logging** debugging option, explicity
triggers all CRUD operations to write to log for debugging purposes. The
250 threads will all write to the logging subsystem. This creates
contention within the logging system.

The fix is to disable table_logging, it is only used for debugging
purposes and should not be enabled as part of a performance test.