WT-14240 Use WT_CONN_READY to determine if the connection is active, not WT_CONN_CLOSING (#11797)
Delay the background migration threads from beginning work until the
connection is fully started, and join the threads earlier in the
connection close process.

The original problem is read-modify-write races when calling the
thread-unsafe `F_SET` macro, which can result in `F_SET` not actually
setting a connection flag. There are two ways that this is triggered in
our live restore tests:

1. On `connection_open` we start up the background migration threads,
they find an empty migration work queue and immediately shut down. This
process can trigger a checkpoint which calls `F_SET(conn,
WT_CONN_CKPT_GATHER)` that can race with any F_SET calls performed when
setting the connection configuration

2. If a `connection_close` is called in the middle of a live restore
clean up the same race and error can arise from `F_SET(conn,
WT_CONN_CKPT_GATHER)` and `F_SET(conn, WT_CONN_CLOSING)` being called in
parallel.

This change does not fix the read-modify-races, instead it reduces the
race window or mitigates the impact of a race. By delaying the startup
of the migration threads until connection configuration is complete we
no longer hit scenario (1) by not calling the `F_SET` operations in
parallel.
Scenario (2) is not fixed, but the chances of it firing are much rarer
since it requires migration to complete at the exact same time as
`connection_close` is called. In comparison scenario (1) is likely to
fire if a connection is ever started when the work queue is empty.