WT-14348 Add arg to decode tool to optionally print extent lists (#11820)
The decode tool currently requires the argument --verbose to dump the
extent list. This also has the affect of dumping all the keys and values
in the blocks. This PR is to add an argument to allow the tool to only
decode the extent lists without dumping the other block data.