WT-13310 Random Cursor (#11791)
This PR fixes the issue where the session RNG returns the same sequence
of numbers.

1. Because the session's `rnd` uses a carefully crafted initial seed to
work best with skiplists, this PR introduces another per-session random
generator for general purposes. To avoid confusion, these generators are
given specific names: `rnd_skiplist` and `rnd_random`.

2. Functions have been renamed for better clarity and readability:
* `__wt_random_init_seed` is renamed to `__wt_session_rng_init_once`,
and its implementation is modified:
* It will initialize the session's RNG at most once (using
`WT_SESSION_FIRST_USE()`).
* It requires `session->id` to be set before the call to this function.
* `rnd_skiplist` is initialized the same way as before, maintaining the
specialized seed.
* `rnd_random` is initialized with a uniquely selected seed to ensure it
never overlaps within the same application.
* `__wt_random_init` is renamed to `__wt_random_init_default` to
highlight that it uses the default seed.
* `__wt_random_init_custom_seed` is renamed to `__wt_random_init_seed`
for conciseness - since the user-supplied seed is always a "custom" one.

3. Removed unnecessary `volatile` keywords.

4. Other cleanup of no longer used code and macros.

The new test `test/csuite/random_session/test_random_session` attempts
to reproduce the bug in two different ways:
1. It opens a session, generates a random number with its RNG and closes
the session (in a loop). In this mode, in fact, the issue doesn't show
up because the session's RNG is only initialized once and the same
session object is getting reused.
2. Rapidly creates a bunch of sessions and generates a random number
with their RNGs. This actually reproduces the issue in the `develop`
branch but this PR fixes it.