WT-14317 Create macro which aborts when conn->close() returns an error code (#11829)
Currently we have BFs that occur due to an EBUSY error code returned
from conn->close(). The ticket aims to add diagnostic information for
when an EBUSY error code happens again. A new macro has been introduced
which will check if the connection is closing and check if the return
value is non-zero. If both conditions are met, we abort in an attempt to
narrow down where the EBUSY could come from. In the pull request I have
identified key areas that the new macro could catch an EBUSY error.