WT-13585 wt_cmp_dir - fix handling WT logs output during check execution (#11833)
Currently, if some log output occurs during the execution of the
wt_cmp_dir script, the script fails because it mixes the output of the
`wt list` internal call and the log output. This PR fixes it by
redirecting `wt list` output to the specified file and redirecting all
the unrelated output during these checks to `/dev/null`.

The original issue doesn't reproduce on the default CI. Please see the
corresponding Jira ticket for testing proofs.