WT-14372 Remove check that prevents evergreen script from running on forked branches (#11843)
The check for `wiredtiger/wiredtiger` in `evg_cfg.py` causes this script
to fail on forks of the wiredtiger repo, as the git remote.origin.url
does not contain the correct string.

This fix removes the check for the specific wiredtiger name, and instead
uses the location of the `test/evergreen/evg_cfg.py` file to find the
wiredtiger root folder.