SLS-1956 Reconciliation performance fixes (#11870)
All credit to Alex G for these!

This PR includes the following:

* Loosened materialisation checks. This is the key part of the branch,
and allows scrub eviction when the materialisation frontier is older
than the page.
* Allow pages to be reconciled before checkpoint visits them
* Stats for tracking large pages. This also extends the set to 1GB
(compared to 100MB in Alex's branch).
* Stat for update restore giving up on a split due to lock contention

Notably, it does NOT include:
* MongoDB table creation workarounds: see SLS-1970 for those. I may
raise a separate Mongo PR today with a big `TODO` comment for this.
* Assertion + crash when reconciling pages over 1GB
* The two-second eviction sleep: This never really appeared to work (at
least from looking at FTDC). It's also a workaround for some weird
observed MongoDB behaviour where we seem to be continuously
checkpointing. This needs separate following up.
* Eviction dirty target and trigger increase. This is of marginal help,
but does make a difference. This is really a longer-term play, since
with precise checkpoint we expect more dirty content to live in the
cache as a matter of course.
* SLS-1988: separate PR.