WT-12244 Add comment on rec_row_zero_len (#11888)
[WT-12244](https://jira.mongodb.org/browse/WT-12244) ticket was to
investigate whether the `__rec_row_zero_len` function can be simplified
by removing a check.

While logically the extra check is not required as
`__wt_txn_tw_start_visible_all` check covers it, it is an early exit to
the function and can allow the method to skip evaluating the more
expensive visible_all check. This pull request adds a comment explaining
this. We can keep the code as is close the ticket as won't do.