WT-14391 Check eviction server is running before opening HS (#11857)
Fix for issue investigated in WT-14376 Data corruption during recovery
as a result of opening History Store.

This moves the initial checks in `__wti_evict_app_assist_worker` to the
calling function `__wt_evict_app_assist_worker_check`. This calling
function contains all the deciding parameters for determining whether
eviction should proceed. With those checks occurring first, it is then
safe to open the history store in `__wti_evict_app_assist_worker`.

Prior to this fix, it was possible to open the history store during
metadata recovery due to the eviction check. As a result, the HS could
have loaded an older checkpoint, before replaying the metadata for the
latest checkpoint on the HS. Then, when the subsequent live checkpoint
is taken the newer checkpoint extent lists are rolled into the older
checkpoint's extent lists which can cause a block merge error due to
incorrect overlaps between extents used.