SLS-2840 Add per-page history tracking (#11882)
Add a debugging feature for tracking per-page statistics, e.g., for
understanding the cache behavior of certain hot pages.

The PR also adds a simple, generic hash map implementation (used here),
which could be hopefully a useful component for future prototyping
and/or for use cases that are not performance critical.