WT-14403 Remove the partial logic and rewrite __live_restore_can_service_read (#11871)
This PR is to remove the partial logic and rewrite the function
`__live_restore_can_service_read`. The main changes include:
1. Rewrite the logic to check varies scenarios during the iteration from
`read_start_bit` to `read_end_bit`.
2. Remove the PARTIAL logic, now the read is either read from source or
dest. The reason behind that is since wiredtiger always read/write one
page at a time, when a read range is partial, it means there's an in
progress background migration on that page but there are no writes on
the page yet, thus read the entire page from source should be safe.