WT-14127 Switch over to V5 toolchain within all supported machines (#11891)
The pull request changes switches all testing machines to use the new V5
toolchain. Here are a few technical changes that were needed to
accommodate this change:
- Remove msan_suppression_wrappers.c because WiredTiger's minimum LLVM
version is now greater than 14 (WT-13103).
- Update clang scan diff to accomodate for a new false positive.
(WT-14298)
- Fix tsan library path for python tests. (WT-14298)
- Update compatibility tests to use mongodbtoolchain_stable_gcc
toolchain.
- Temporarily comment out PRELOAD of tcmalloc (WT-14389).
- Remove S3 tests from ARMV9 due to incompatibility issues.
- Remove linting from ubuntu-arm64 machines. We should have one lint
task in a patch build to make formatting consistent.
- Use the v4 mongodbtoolchain for ARMV9 machines due to workgen tests
failing. (WT-14461 to investigate)

It seems WT-14389 is merging very soon, I will update the PR when that
is merged in.

---------

Co-authored-by: Ivan Kochin <ivan.kochin@mongodb.com>