WT-14561 Build fixes for GCC 15 (#11903)
The `-Warray-bounds` thing appears to have been fixed in GCC 15, so I
removed the message.

There's also a new "unterminated string" warning when a string doesn't
have a terminating null. test/checkpoint hit this, but it's OK because
we don't actually use it as a string, just an array of characters. The
alternative to removing the digit count would have been to mark it with
`__attribute((nonstring))` or however it's spelled.