WT-10792 csuite tests take absolute and relative path home dir (#11831)
**tl;dr** - `random_abort`, `timestamp_abort`, `schema_abort`,
`random_directio` and other csuite tests are now able to take absolute
and relative path arguments when specifying a database home directory
with the `-h` flag, and leave no test artefacts after the test
concludes.

The original (undocumented) assumption was that anything passed to these
test scripts with the `-h` argument is a single directory, not a path.
The tests then simply append test artefact names (i.e. "`SAVE`") when
copying the db, and prepend "`../`" when they need to find related
directories. The tests expect to be given a `WT_TEST` directory that
contains a database home directory named `WT_HOME` when the `-h` flag is
specified. The changes in this PR makes it so that the tests are not
only able to take single directories, but also absolute and relative
paths to these `WT_TEST` directories, and are no longer dependent on
hardcoded paths to locate test artefacts. In the `test/utility/misc.c`
files, the `testutil_copy_data`, `testutil_copy_data_opt` and
`testutil_clean_test_artifacts` no longer rely on hardcoded paths to
generate and cleanup test artefacts, but instead generate them in the
provided `WT_TEST` directory.

The `random_abort`, `timestamp_abort`, `schema_abort` and
`random_directio` tests exited normally with these changes under the
following cases:
- db provided as a relative path from within test dir
- db provided as an absolute path from within test dir
- db provided as a single dir name from within test dir
- db provided as a single dir name from outside test dir
- db provided as a relative path from outside test dir
- db provided as an absolute path from outside test dir

Some other notes to consider:
- Both `normalized_pos`, `incr_backup`,
`wt8246_compact_rts_data_correctness` and `wt10461_skip_list_stress`
unnecessarily call `testutil_clean_test_artifacts`, so it has been
removed. If future code changes in these tests generate artefacts, it is
logical to ensure their cleanup is explicitly handled.

Upon investigating the issues related to this ticket, other issues/bugs
were discovered and tickets have been created for them:
- [WT-14283](https://jira.mongodb.org/browse/WT-14283) - When a user
specifies the `-v` flag to perform a verify only operation with the
csuite tests, this operation is not performed
under `read_only` circumstances and the originally provided database is
not preserved (it is deleted).
- [WT-14288](https://jira.mongodb.org/browse/WT-14288) - When a user
specifies a database home directory using the `-h` flag and adds
a `/` at the end of the path, the test falls into an issue where it
recursively makes endless copies of test artefacts before eventually
crashing and eating up GBs of disc space. **Note**: this PR addresses
this issue for all tests except `random_directio`.
- [WT-14355](https://jira.mongodb.org/browse/WT-14355) - leaves nested
test artefacts when the preserve `-p` flag is not set.